Move Messaging component to common. Display in holiday modal so admins can see and send replies.
