+++ src/components/index.js
@@ -24,3 +24,4 @@ export { default as DataTable } from './DataTable';
 export { default as FullPageLoader } from './FullPageLoader';
 export { default as TeamOverview } from './TeamOverview';
 export { default as LoginLoader } from './LoginLoader';
+export { default as Messaging } from './Messaging';
